* logger.c (klog_vsyslog): Convert pid_t to long for printing
* alt_prof.c (kadm5_get_config_params): Fix bug in direction of
test while walking through whitespace.  Thanks to Matt Crawford.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@11951 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/kadm5/logger.c
@@ -853,8 +853,8 @@ klog_vsyslog(priority, format, arglist)
     cp += 15;
 #endif	/* HAVE_STRFTIME */
 #ifdef VERBOSE_LOGS
-    sprintf(cp, " %s %s[%d](%s): ", 
-	    log_control.log_hostname, log_control.log_whoami, getpid(),
+    sprintf(cp, " %s %s[%ld](%s): ",
+	    log_control.log_hostname, log_control.log_whoami, (long) getpid(),
 	    severity2string(priority));
 #else
     sprintf(cp, " ");
